@@ -2,19 +2,23 @@
   "name": "terra-trade",
   "version": "0.1.0",
   "private": true,
-"scripts": {
-  "build": "dotenv -e .env.production next build",
-  "dev": "dotenv -e .env.development next dev",
-  "start": "next start"
-}
-,
-
+  "scripts": {
+    "build": "dotenv -e .env.production next build",
+    "dev": "dotenv -e .env.development next dev",
+    "start": "next start"
+  },
   "dependencies": {
     "@metamask/sdk-react": "^0.14.1",
     "@supabase/auth-helpers-react": "^0.5.0",
+    "@supabase/ssr": "^0.5.2",
+    "@types/next": "^9.0.0",
+    "dotenv": "^16.4.7",
+    "encoding": "^0.1.13",
     "ethers": "^6.13.2",
-    "next": "13.5.5",
+    "next": "^15.0.3",
     "next-auth": "^4.24.7",
+    "next-compose-plugins": "^2.2.1",
+    "next-pwa": "^5.6.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "styled-components": "^6.1.11",
@@ -26,6 +30,7 @@
     "@types/node": "^20",
     "@types/react": "^18.2.0",
     "@types/react-dom": "^18.2.0",
+    "dotenv-cli": "^7.4.4",
     "eslint": "^8",
     "eslint-config-next": "^13.5.5",
     "tailwindcss": "^3.4.4",
